==> backend/notebooks/05. Creating Sample Data.ipynb <==
# **Creating Sample Data**
While I'm quickly prototyping functionalities, I don't want to use my entire `conversations.json` - it's too large. Instead, I'll create a sample of it!

# Change the working directory 
%cd ..

# Enable the autoreload extension, which will automatically load in new code as it's written
%load_ext autoreload
%autoreload 2

# General imports 
import random
import json
from pathlib import Path

First: declare some parameters!

# Declare the parameters
conversations_json_path = Path("D:/data/programming/chatgpt-export-explorer/api/data/export/conversations.json")
output_path = "data/sample_conversations.json"
n_conversations_in_sample = 100
random_seed = 42

Next: I'll create the sample. 

# Seed the random number generator
random.seed(random_seed)

# Load the conversations
with open(conversations_json_path, "r") as f:
    conversations = json.load(f)

# Sample the conversations
sampled_conversations = random.sample(conversations, n_conversations_in_sample)

# Save the sampled conversations
output_path = Path(output_path)
output_path.parent.mkdir(parents=True, exist_ok=True)
with open(output_path, "w") as f:
    json.dump(sampled_conversations, f)

# Print the output path
print(
    f"Created sample with {len(sampled_conversations)} conversations at {output_path}"
)
